Update plotting tests
